updated naming and moved kernel up
