+++ attn_gym/sparse/indexer/impl/cute/__init__.py
@@ -1,5 +1,5 @@
 """CuTe DSL (SM100) backend for the multi-head weighted ReLU Top-K indexer."""
 
-from .prefill import index
+from .impl import index
 
 __all__ = ["index"]
